mv Sequential, ParameterList, LayerList from paddle.fluid.dygraph.container to paddle.nn.container
