+++ python/paddle/fluid/tests/unittests/test_imperative_container_layerlist.py
@@ -31,11 +31,6 @@ class MyLayer(fluid.Layer):
 
 
 class TestImperativeContainer(unittest.TestCase):
-    def fluid_dygraph_list(self):
-        return fluid.dygraph.LayerList(
-            [fluid.dygraph.Linear(2**i, 2 ** (i + 1)) for i in range(6)]
-        )
-
     def paddle_imperative_list(self):
         return paddle.nn.LayerList(
             [fluid.dygraph.Linear(2**i, 2 ** (i + 1)) for i in range(6)]
@@ -45,11 +40,7 @@ class TestImperativeContainer(unittest.TestCase):
         data_np = np.random.uniform(-1, 1, [5, 1]).astype('float32')
         with fluid.dygraph.guard():
             x = fluid.dygraph.to_variable(data_np)
-            layerlist = (
-                self.fluid_dygraph_list()
-                if use_fluid_api
-                else self.paddle_imperative_list()
-            )
+            layerlist = self.paddle_imperative_list()
             size = len(layerlist)
 
             model = MyLayer(layerlist)
