+++ python/paddle/fluid/tests/unittests/test_imperative_container_parameterlist.py
@@ -24,17 +24,7 @@ class MyLayer(fluid.Layer):
     def __init__(self, num_stacked_param, use_fluid_api):
         super().__init__()
         # create ParameterList with iterable Parameters
-        self.params = (
-            self.fluid_dygraph_ParameterList(num_stacked_param)
-            if use_fluid_api
-            else self.paddle_imperative_ParameterList(num_stacked_param)
-        )
-
-    def fluid_dygraph_ParameterList(self, num_stacked_param):
-        return fluid.dygraph.ParameterList(
-            [fluid.layers.create_parameter(shape=[2, 2], dtype='float32')]
-            * num_stacked_param
-        )
+        self.params = self.paddle_imperative_ParameterList(num_stacked_param)
 
     def paddle_imperative_ParameterList(self, num_stacked_param):
         return paddle.nn.ParameterList(
